Move data classes to seperate package
