+++ core/src/main/kotlin/com/malinskiy/marathon/report/junit/model/TestCaseData.kt
@@ -1,0 +1,9 @@
+package com.malinskiy.marathon.report.junit.model
+
+data class TestCaseData(
+    val classname: String,
+    val name: String,
+    val time: String,
+    val skipped: String,
+    val failure: String
+)
