+++ core/src/main/kotlin/com/malinskiy/marathon/report/junit/model/TestSuiteData.kt
@@ -1,0 +1,11 @@
+package com.malinskiy.marathon.report.junit.model
+
+data class TestSuiteData(
+    val name: String,
+    val tests: Int,
+    val failures: Int,
+    val errors: Int,
+    val skipped: Int,
+    val time: String,
+    val timeStamp: String
+)
